refactor(otel): make OTLP tracing link-time opt-in
Split internal/otel into a lightweight facade (Init/Shutdown/Register)
and a public go-micro.dev/v6/otel backend that mounts the OTLP trace
exporter. Core packages keep calling the facade; with no backend linked
both calls no-op and the otel SDK, otlptracehttp, and grpc stay out of
the build. cmd/micro blank-imports the backend to keep CLI tracing.

Consumer probe: 494 -> 385 packages, stripped binary 20.1M -> 17.1M,
66 grpc packages -> 0, removing the largest hard import from core.
