+++ cmd/micro/main.go
@@ -4,7 +4,11 @@ import (
 	"embed"
 
 	"go-micro.dev/v6/cmd"
-	"go-micro.dev/v6/internal/otel"
+	builtin "go-micro.dev/v6/internal/otel"
+
+	// Link the OTLP trace exporter so CLI-hosted services/agents/flows emit
+	// spans when OTEL_EXPORTER_OTLP_ENDPOINT is set. Library users omit it.
+	_ "go-micro.dev/v6/otel"
 
 	// Link every plugin so CLI flag selection (--registry etcd, --broker nats,
 	// --profile nats, ...) keeps working; library users omit this import.
@@ -36,7 +40,7 @@ func init() {
 }
 
 func main() {
-	otel.Init()
+	builtin.Init()
 	_ = cmd.Init(
 		cmd.Name("micro"),
 		cmd.Version(version),
